Address code review feedback
- Remove plan.md from repo (working document, not project code)
- Extract duplicated isDark media query into usePrefersDark hook
- Remove unnecessary "use client" directive from footer component
- Use Next.js Link instead of <a> for internal routes in legal pages
