+++ frontend/src/app/(landing)/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useEffect, useState } from "react";
+import { usePrefersDark } from "@/hooks/use-prefers-dark";
 import HeroSection from "@/components/landing/hero-section";
 import HowItWorksSection from "@/components/landing/how-it-works-section";
 import StorySection from "@/components/landing/story-section";
@@ -8,15 +8,7 @@ import CtaSection from "@/components/landing/cta-section";
 import Footer from "@/components/landing/footer";
 
 export default function LandingPage() {
-  const [isDark, setIsDark] = useState(true);
-
-  useEffect(() => {
-    const mq = window.matchMedia("(prefers-color-scheme: dark)");
-    const update = () => setIsDark(mq.matches);
-    update();
-    mq.addEventListener("change", update);
-    return () => mq.removeEventListener("change", update);
-  }, []);
+  const isDark = usePrefersDark();
 
   return (
     <div className="relative">
