+++ frontend/src/components/landing/footer.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import Link from "next/link";
 
 interface FooterProps {
